+++ src/server/devices/zigbee/index.ts
@@ -14,6 +14,7 @@ export * from './zigbeeIlluLampe';
 export * from './zigbeeIlluLedRGBCCT';
 export * from './zigbeeIlluShutter';
 export * from './zigbeeLinkindLedRgbCct';
+export * from './zigbeeOsramDimmer';
 export * from './zigbeeSMaBiTMagnetContact';
 export * from './zigbeeSonoffMotion';
 export * from './zigbeeSonoffTemp';