+++ CHANGELOG.md
@@ -4,11 +4,14 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
--->
+--> 
+
+## **WORK IN PROGRESS**
+  * (nockiro) Fixes for Osram dimmer
+  
 ## 1.1.43 (2022-10-15)
 
 * (nockiro) Add Osram dimmer device
-*
 
 ## 1.1.42 (2022-10-15)