+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 1.1.45 (2022-10-15)
 * (nockiro) Fixes/Adjustments for PostgreSQL to create schema/tables correctly
 
 ## 1.1.44 (2022-10-15)